Admin can to remove user msgs

+++ api/review/review.controller.js
@@ -3,6 +3,8 @@ import { socketService } from '../../services/socket.service.js'
 import { userService } from '../user/user.service.js'
 import { authService } from '../auth/auth.service.js'
 import { reviewService } from './review.service.js'
+import { toyService } from '../toy/toy.service.js'
+
 
 export async function getReviews(req, res) {
 	try {
@@ -19,6 +21,11 @@ export async function deleteReview(req, res) {
 	const { id: reviewId } = req.params
 
 	try {
+		const review = await reviewService.getById(reviewId)
+        if (!loggedinUser.isAdmin && review.userId !== loggedinUser._id) {
+            return res.status(403).send({ err: 'Not authorized to delete this review' })
+        }
+		
 		const deletedCount = await reviewService.remove(reviewId)
 		if (deletedCount === 1) {
 			socketService.broadcast({ type: 'review-removed', data: reviewId, userId: loggedinUser._id })
